+++ package.json
@@ -12,8 +12,8 @@
   },
   "devDependencies": {
     "browser-sync": "2.26.14",
-    "eslint": "7.18.0",
-    "eslint-config-htmlacademy": "4.0.0"
+    "eslint": "^7.18.0",
+    "eslint-config-htmlacademy": "4.0.1"
   },
   "license": "MIT",
   "dependencies": {},
